cli: keep resource webapi wrappers together
