+++ flow360/component/simulation/web/asset_webapi.py
@@ -1,49 +1,94 @@
 """
-Thin asset web API wrappers.
+Thin V2 resource web API wrappers.
 """
 
 from __future__ import annotations
 
+import json
+
+from flow360.cloud.rest_api import RestApi
 from flow360.component.interfaces import (
     CaseInterfaceV2,
+    DraftInterface,
     GeometryInterface,
     SurfaceMeshInterfaceV2,
     VolumeMeshInterfaceV2,
 )
-from flow360.component.simulation.web.resource_webapi import ResourceWebApi
 
 
-class AssetWebApi(ResourceWebApi):
-    """Thin wrapper around a single asset endpoint."""
+class ResourceWebApi:
+    """Thin wrapper around a single Flow360 resource endpoint."""
 
-    def __init__(self, interface, asset_id: str):
-        self.asset_id = asset_id
-        super().__init__(interface, asset_id)
+    def __init__(self, interface, resource_id: str):
+        self.resource_id = resource_id
+        self._api = RestApi(interface.endpoint, id=resource_id)
+
+    @staticmethod
+    def _unwrap_data(response):
+        """Return response data when REST responses use a top-level data envelope."""
+        if isinstance(response, dict) and "data" in response:
+            return response["data"]
+        return response
+
+    def get_info(self):
+        """Fetch resource metadata."""
+        return self._unwrap_data(self._api.get())
+
+    def get_simulation_json(self):
+        """Fetch the resource simulation JSON payload."""
+        response = self._unwrap_data(
+            self._api.get(method="simulation/file", params={"type": "simulation"})
+        )
+        if isinstance(response, dict) and "simulationJson" in response:
+            response = response["simulationJson"]
+        if isinstance(response, str):
+            return json.loads(response)
+        return response
+
+    def get(
+        self, path=None, method=None, json=None, params=None
+    ):  # pylint: disable=redefined-outer-name
+        """Delegate specialized GET calls to the underlying REST API."""
+        return self._api.get(path=path, method=method, json=json, params=params)
 
 
-class GeometryWebApi(AssetWebApi):
+class GeometryWebApi(ResourceWebApi):
     """Thin geometry web API wrapper."""
 
     def __init__(self, asset_id: str):
         super().__init__(GeometryInterface, asset_id)
 
 
-class SurfaceMeshWebApi(AssetWebApi):
+class SurfaceMeshWebApi(ResourceWebApi):
     """Thin surface mesh web API wrapper."""
 
     def __init__(self, asset_id: str):
         super().__init__(SurfaceMeshInterfaceV2, asset_id)
 
 
-class VolumeMeshWebApi(AssetWebApi):
+class VolumeMeshWebApi(ResourceWebApi):
     """Thin volume mesh web API wrapper."""
 
     def __init__(self, asset_id: str):
         super().__init__(VolumeMeshInterfaceV2, asset_id)
 
 
-class CaseWebApi(AssetWebApi):
+class CaseWebApi(ResourceWebApi):
     """Thin case web API wrapper."""
 
     def __init__(self, asset_id: str):
         super().__init__(CaseInterfaceV2, asset_id)
+
+
+class DraftWebApi(ResourceWebApi):
+    """Thin draft web API wrapper."""
+
+    def __init__(self, draft_id: str):
+        super().__init__(DraftInterface, draft_id)
+
+    @classmethod
+    def list_records(cls, project_id: str):
+        """List draft records for a project."""
+        api = RestApi(DraftInterface.endpoint)
+        response = api.get(params={"projectId": project_id})
+        return response.get("records", [])
